Allow (srfi 1) to fill in for (scheme list) when latter unavailable.
Add a couple of cond-expands to allow (srfi 1) to fill in for (scheme
list).  This seems to be currently required in order to get
(slib xml-parse) working with Kawa. There are other places that could
use a simliar change, but I am holding off for now in case I'm
overlooking something.